server/requirements.txt
Django==4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.2
pytest==7.0.0
pytest-django==4.5.2
django_extensions==3.1.5
marshmallow==3.14.1
bleach==4.1.0
factory_boy==3.2.1
